GUI/int when posible

+++ resources/shaders/DepthRTT.vert
@@ -1,11 +1,12 @@
-#version 330 core
+#version 450 core
 
 // Input vertex data, different for all executions of this shader.
-layout(location = 0) in vec3 vertexPosition_modelspace;
+layout(location = 0) in ivec3 vertexPosition_modelspace;
 
 // Values that stay constant for the whole mesh.
 uniform mat4 depthMVP;
 
-void main(){
-	gl_Position =  depthMVP * vec4(vertexPosition_modelspace,1);
+void
+main() {
+    gl_Position = depthMVP * vec4(vertexPosition_modelspace, 1);
 }
